backend/requirements.txt
fastapi
uvicorn
pymongo
bcrypt
PyJWT
